+++ src/interfaces/node.h
@@ -113,8 +113,7 @@ public:
     virtual bool getBanned(BanTables &banmap) = 0;
 
     //! Ban node.
-    virtual bool ban(const CNetAddr &net_addr, BanReason reason,
-                     int64_t ban_time_offset) = 0;
+    virtual bool ban(const CNetAddr &net_addr, int64_t ban_time_offset) = 0;
 
     //! Unban node.
     virtual bool unban(const CSubNet &ip) = 0;